+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Change Log
 - Refine the return of LLM[`#83 <https://github.com/defeat-beta/defeatbeta-api/issues/83>`_]
 - Support Industry PE[`#67 <https://github.com/defeat-beta/defeatbeta-api/issues/67>`_]
 - IO Error: SSL connection failed error for HTTP HEAD to xxxx[`#87 <https://github.com/defeat-beta/defeatbeta-api/issues/87>`_]
+- Optimize the perf of Industry PE[`#88 <https://github.com/defeat-beta/defeatbeta-api/issues/88>`_]
 
 0.0.23
 -------